add settings conception for key storage
