+++ src/lib.rs
@@ -46,6 +46,10 @@ mod view_state;
 mod test_utils;
 mod linux_key_storage;
 
+mod file_key_storage;
+mod security_framework_key_storage;
+mod settings;
+
 pub use app::Damus;
 pub use error::Error;
 pub use profile::DisplayName;
